pull benchmark code out into its own module
